MNIST CNN example.
CNN for the MNIST dataset was added. However, the current version
does't have pooling layers. So we will update this example once pooling
layers (average, and max pooing) are ready.

Since, the GPU version of the conv and pooing layer not yet implemented
exceptions will be raised if you use -c gpu flag.