Add bias to the conv layer
`conv2d` layer was updated and now it takes a bais vector as an argument.
In order to implement this we introduced a new node
(`Conv2dBackwardBias`) to our auto diff framework. In addition to that
we have updated `compute` and `backward` method of the `Conv2dOp` node.

Also, `mnist_cnn.py` was updated to utilize new changes we introduced to
the `conv2d` layer.